Changes to weapon system. Started UI.
Made it so that the weapon system is way more modular, making it easier to add nem weapons.
Also started the UI, which already has a working ammo counter.
The ammo system has a bug where reloading when the ammo reserve is smaller than the gun's magazine size will result in all of the reserve going to the magazine, ususally exceeding it's capacity.